Refactor CLI logger and util functions
