+++ pythx/cli/logger.py
@@ -1,0 +1,9 @@
+import logging
+from os import environ
+
+if environ.get("PYTHX_DEBUG") is not None:
+    logging.basicConfig(level=logging.DEBUG)
+else:
+    logging.basicConfig(level=logging.ERROR)
+
+LOGGER = logging.getLogger("pythx-cli")
