+++ pythx/models/request/base.py
@@ -5,7 +5,6 @@ import json
 import logging
 
 import jsonschema
-
 from pythx.models.exceptions import RequestValidationError
 
 LOGGER = logging.getLogger(__name__)
