+++ pythx/models/response/base.py
@@ -5,7 +5,6 @@ import json
 import logging
 
 import jsonschema
-
 from pythx.models.exceptions import ResponseValidationError
 
 LOGGER = logging.getLogger(__name__)
