+++ tests/common.py
@@ -1,7 +1,6 @@
 from datetime import datetime
 
 import dateutil.parser
-
 from pythx.models.request import (
     AnalysisListRequest,
     AnalysisStatusRequest,
