+++ tests/test_analysis_list_request.py
@@ -2,7 +2,6 @@ import json
 
 import dateutil.parser
 import pytest
-
 from pythx.models.exceptions import RequestValidationError
 from pythx.models.request import AnalysisListRequest
