+++ tests/test_analysis_list_response.py
@@ -2,7 +2,6 @@ import json
 from copy import deepcopy
 
 import pytest
-
 from pythx.models.exceptions import ResponseValidationError
 from pythx.models.response import AnalysisListResponse
 from pythx.models.response.analysis import Analysis
