+++ tests/test_auth_logout_response.py
@@ -1,7 +1,6 @@
 import json
 
 import pytest
-
 from pythx.models.exceptions import ResponseValidationError
 from pythx.models.response import AuthLogoutResponse
