+++ tests/test_client_tool_name_middleware.py
@@ -1,5 +1,4 @@
 import pytest
-
 from pythx.middleware.toolname import ClientToolNameMiddleware
 
 from . import common as testdata
