+++ tests/test_config.py
@@ -1,5 +1,4 @@
 import pytest
-
 from pythx import config
 from pythx.conf.base import PythXConfig
