+++ tests/test_detected_issues_request.py
@@ -1,7 +1,6 @@
 import json
 
 import pytest
-
 from pythx.models.exceptions import RequestValidationError
 from pythx.models.request import DetectedIssuesRequest
