+++ tests/test_model_validator_errors.py
@@ -1,5 +1,4 @@
 import pytest
-
 from pythx.models import request as reqmodels
 from pythx.models import response as respmodels
 from pythx.models.exceptions import RequestValidationError, ResponseValidationError
