+++ tests/test_version_request.py
@@ -1,7 +1,6 @@
 import json
 
 import pytest
-
 from pythx.models.request import VersionRequest
 
 from . import common as testdata
